requirements.txt
fastapi
uvicorn
pandas
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy